# speed up mosaic with StackView