--- a/Project.toml
+++ b/Project.toml
@@ -7,11 +7,13 @@ version = "0.3.0"
 MappedArrays = "dbb5928d-eab1-5f90-85c2-b9b0edb7c900"
 OffsetArrays = "6fe1bfb0-de20-5000-8ca7-80f57d26f881"
 PaddedViews = "5432bcbf-9aad-5242-b902-cca2824c8663"
+StackViews = "cae243ae-269e-4f55-b966-ac2d0dc13c15"
 
 [compat]
 MappedArrays = "0.2, 0.3"
 OffsetArrays = "0.10, 0.11, 1"
 PaddedViews = "0.5.8"
+StackViews = "0.1.1"
 julia = "1"
 
 [extras]