By applying `StackView` to `_padded_cat`, unnecessary memory allocation is avoided and thus boost the performance

```julia
using ImageCore
using TestImages
using ColorVectorSpace
using MosaicViews

img = testimage("mri-stack");
framestack = [img[:,:,i] for i in 1:size(img, 3)];
@btime mosaic($framestack);
# master: 793.137 μs (1067 allocations: 1.18 MiB)
# PR: 18.876 μs (477 allocations: 52.27 KiB)

# heterogeneous case
A = testimage("toucan");
B = testimage("moon");
@btime mosaic($A, $B);
# master: 519.649 μs (48 allocations: 514.50 KiB)
# PR: 1.381 μs (24 allocations: 1.62 KiB)

framestack = [rand(32, 32) for i in 1:1000];
@btime mosaic($framestack);
# master: 20.816 ms (10001 allocations: 21.66 MiB)
# PR: 9.010 ms (3511 allocations: 6.94 MiB)
```

This is not a thorough optimization, I only meant to fixes the unnecessary array allocation part in this PR.

~Unfortunately, I'm not familiar with the internal inference mechanism so it actually breaks the inferrability improvements introduced in #28 and thus the inferred tests. (cc: @timholy)~

~Eventually, I'd like to make `_padded_cat` a standard tool so that other packages could use this. Perhaps a new package `StackViews.jl` would be sufficient, maybe, since `_padded_cat = paddedviews |> StackView` in its simplest form.~