@johnnychen94 2021-04-08
I've moved the `StackView` thing to a separate package https://github.com/JuliaArrays/StackViews.jl
WIll continue to work on this when that's registered. (3 days later)